pyproject.toml
[project]
name = "cogneuro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "einops>=0.8.2",
    "ipykernel>=7.2.0",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
]
